Completed the new layer system for rendering. This is how we will support things like sky boxes. This could also be extended to things like first person weapons if desired. At the moment, there is no support for culling objects based on layers, but that should be implemented eventually.
The way it works is that instead of attaching only a single camera to a viewport, you can now attach any number, only now each camera is associated with a layer, with each layer only supporting a single camera. Layers with smaller index's are rendering underneath everything else. As an example, you could draw a sky box on layer -1 and the main world on layer 0 and maybe the first person weapon on layer +1.