Improve the way the position gizmo is rendered.
The previous system loaded an arrow model from a file, but now we just build the arrow vertex data procedurally. The main benefit here is that we can define the line part of the arrow as a line rather that solid geometry. This makes it look much, much nicer in the editor.