Add some attributes to EditorMetadataComponent to allow a custom model to be specified.
As of this commit nothing actually uses this. What this is useful for is to allow something like a sprite to be used in an editor instead of having to add a special node with a ModelComponent. This system should be a bit simpler than that.

One of the properties is a custom transformation. When enabled, the scene node transformation will be ignored and the custom transformation used instead. Really useful for doing camera-facing billboards/sprites.