Add events to Scene that are called when a component is added/removed from a scene node.
This commit also updates the appropriate managers when the state of the components of a scene node changes. This will eventually mean the end to the annoying refreshing system current in place.

Another reason this is needed is that it will simplify a few things with the scene state stack and scene node deserialization.