Improve the culling manager interface.
This change makes it so that instead of using a single method for updating transformations, a separate method is used for each type of object. This is required to support some of the recent changes to the way changes to scene node components are handled by the owner scene.

Another improvement with this commit is that a custom data pointer is no longer set on scene nodes. This system was just stupid. Instead, metadata is mapped to a scene object via a GTCore::Map container. It's slower, but nicer, cleaner and less error prone.

This commit includes dead code, which will be removed in the next commits.