Remove the type ID system from SceneNode.
This feature was intended to make it easier for a client application to associate a type to a scene node. The problem is that this goes against the whole point of having a component system where there isn't really the idea of a "type" of scene nodes. Removing this just keeps everything that tiny bit simpler.

The serialized formated has changed with this commit.