Add prototypes for Gaussian blur shader generators.
This new system is needed in order to make Gaussian blurring a little less cumbersome and generalized. Also, it'll make it easier to allow customizations. For example, the blur kernels can potentially be reduced or increased for individual shadow-casting lights.

Also, this is going to be GLSL 110 compliant which is going to be a partial fix for the OpenGL 2.1 renderer.