Remove some unneeded model files.
This is part of an ongoing effort to reduce the size of the engine
distribution. The idea is to keep the engine distribution as small and
simple as possible and to later have a sort of "asset pack" that people
can download separately if they so desire. I want to push the small size
and simplicity of the engine as a "feature".