Introduce the the event filtering system.
This system is designed to make it easy to change the value of input
parameters of events and prevent events from getting dispatched. The
main user of this system at the moment is the scene editor which uses it
to modify the input parameters of mouse events when a scene is being
played in-editor.