Work on adding support for OnSerialize() and OnDeserialize() for script components.
This is work towards addressing issue #88.

All that's really left is calling OnDeserialize() properly, but the issue is
that at the point at which it should be called, the node has not yet been
added to the scene, which in turn means the scripting environment is not
available to the component.