Initial work on moving game event handling over to the new game state manager system.
The idea behind the new game state management system is to have a central
manager for controlling the game state. This will make it easier to do things
such as saving the state to a file (save games) and loading it back up again
(load games).

This is very early work - the API will change going forward.