Add DefaultGameStateManager.
This game state manager is what is used by the Sandbox and should be able to
be used by most games without any significant modification. By moving this out
of the sandbox and into the engine, it is now quite a bit easier to quickly
build executables for things like prototypes or whatnot.