Begin work on the new handle based API for the rendering system.
A handle is a pointer-sized integer variable that holds a pointer to an
object. The advantage to this is that it abstracts the implementation
details which in turn makes the implementation simpler since we no longer
need to use a base class for each and every rendering object (buffers,
shaders, etc.) For example, when using Direct3D, a handle will usually
reference the relevant Direct3D object directly. Previously, it needed
to be wrapped around a separate class.