Add the notion of a low-priority base directory to the file system.
A low priority base directory is intended to be used for base directories
that should always be checked last when looking for files. This is used by
the engine context for the directory where the executable is located. We
want the engine to look in the executable directory for files, but not at
a high priority level, since the application itself will likely know what
the highest priority base directories need to be.