Remove a lot of the old code for window management.
The new policy for this library is that window management is done by the host
application rather than the library. The reason for this change is that
creating a generalized cross-platform interface for the various windowing
systems out there is too complex for this library. Indeed, there are other
libraries out there that can do a much better job.