Add support for loading asset metadata from a file.
Asset metadata is always loaded before the asset file itself. In addition, the metadata will be loaded even if the main file doesn't exist. This is because it is possible that an asset can be fully defined by metadata, in which case it is valid for the base asset file to not exist.